requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.4.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
paypalrestsdk==1.13.3
stripe==7.8.0
sendgrid==6.11.0
django-email-verification==0.3.4